infra/graphiti/pyproject.toml
[project]
name = "graphiti"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "graphiti-core[falkordb]>=0.28.2",
    "python-dotenv>=1.2.2",
]
